remove update/delete buttons if not user's event

+++ event.php
@@ -61,14 +61,20 @@ session_start();
 					<div class="left rounded-outline">
 						<div class="description">
 							' . $result['description'] . '
-						</div>
-						<a style="float:right" id="new-event-btn" class="btn-blue-muted m-2" href="delete.php?id=' . $result['event_id'] . '">Delete</a>
-						<a style="float:right" id="new-event-btn" class="btn-blue-muted m-2" href="update-task.php?id=' . $result['event_id'] . '">Update</a>
-					</div>
+						</div>';
+
+	if ($result['organizer'] == $_SESSION['logged_in']) {
+		echo '
+							<a style="float:right" id="new-event-btn" class="btn-blue-muted m-2" href="delete.php?id=' . $result['event_id'] . '">Delete</a>
+						<a style="float:right" id="new-event-btn" class="btn-blue-muted m-2" href="update-task.php?id=' . $result['event_id'] . '">Update</a>';
+	}
+
+
+	echo '</div>
 		
 					<div class="right rounded-outline">
 						<div class="row">
-							Organizer: <a href="profile.php?username=' . $result['organizer'] . '">' . $result['organizer'] . '</a><br>
+							Host:  <a href="profile.php?username=' . $result['organizer'] . '">@' . $result['organizer'] . '</a><br>
 							Date: ' . $result['date'] . ' <br>
 							Time ' . $result['time'] . '
 						</div>
